+++ package.json
@@ -13,8 +13,6 @@
     "@babel/core": "7.4.4",
     "@babel/preset-env": "7.4.4",
     "@babel/preset-react": "7.0.0",
-    "@types/jest": "24.0.14",
-    "@types/react": "16.8.19",
     "axios-mock-adapter": "1.16.0",
     "babel-jest": "24.8.0",
     "babel-loader": "8.0.5",
@@ -26,10 +24,6 @@
     "eslint-plugin-react": "7.12.4",
     "jest": "24.8.0",
     "react-test-renderer": "16.8.6",
-    "ts-jest": "24.0.2",
-    "ts-loader": "6.0.2",
-    "typescript": "3.5.1",
-    "utility-types": "3.7.0",
     "webpack": "4.30.0",
     "webpack-cli": "3.3.1",
     "webpack-dev-server": "3.3.1"
@@ -46,7 +40,6 @@
     "react-redux": "7.0.3",
     "recompose": "0.30.0",
     "redux": "4.0.1",
-    "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "2.3.0"
   }
 }
