+++ src/components/app/app.test.jsx
@@ -1,9 +1,24 @@
-import * as React from 'react';
+import React from 'react';
 import renderer from 'react-test-renderer';
-import MainScreen from './main-screen';
+import App from './app';
+import {createStore} from 'redux';
+import {initialState, reducer} from '../../reducer';
+import {Provider} from 'react-redux';
+
+const createMockStore = (state = initialState) => createStore(reducer, state);
+
+export const MockProvider = ({state = initialState, children} = {}) => (
+  <Provider store={createMockStore(state)}>
+    {children}
+  </Provider>
+);
 
 const mockData = [
   {
+    city: {
+      name: `Paris`,
+      coords: [1, 3],
+    },
     title: `Beautiful, luxurious apartment at great location`,
     type: `Apartment`,
     coords: [52.3909553943508, 4.85309666406198],
@@ -14,6 +29,10 @@ const mockData = [
     isPremium: true
   },
   {
+    city: {
+      name: `Amsterdam`,
+      coords: [48.8351, 2.3425],
+    },
     title: `Wood and stone place`,
     type: `Private room`,
     coords: [52.369553943508, 4.85309666406198],
@@ -24,6 +43,10 @@ const mockData = [
     isPremium: false
   },
   {
+    city: {
+      name: `Cologne`,
+      coords: [150, 2],
+    },
     title: `Canal View Prinsengracht`,
     type: `Apartment`,
     coords: [52.3909553943508, 4.929309666406198],
@@ -34,6 +57,10 @@ const mockData = [
     isPremium: false
   },
   {
+    city: {
+      name: `Berlin`,
+      coords: [11, 16],
+    },
     title: `Nice, cozy, warm big bed apartment`,
     type: `Apartment`,
     coords: [52.3809553943508, 4.939309666406198],
@@ -42,14 +69,21 @@ const mockData = [
     rate: 100,
     isBookmarked: false,
     isPremium: true
+  },
+  {
+    city: {
+      name: `Cologne`,
+      coords: [150, 2],
+    },
+    title: `Canal View Prinsengracht`,
+    type: `Apartment`,
+    coords: [52.3909553943508, 4.929309666406198],
+    image: `img/apartment-02.jpg`,
+    price: `132`,
+    rate: 80,
+    isBookmarked: true,
+    isPremium: false
   }
-];
-
-const citiesMock = [
-  `Paris`,
-  `Brussels`,
-  `Amsterdam`,
-  `Cologne`
 ];
 
 const fixContainerLeafletTest = () => {
@@ -60,14 +94,13 @@ const fixContainerLeafletTest = () => {
 
 fixContainerLeafletTest();
 
-it(`Main screen correctly renders`, () => {
+it(`App correctly renders`, () => {
   const tree = renderer
-    .create(<MainScreen
-      offers={mockData}
-      onCityClick={jest.fn()}
-      city={`Paris`}
-      cities={citiesMock}
-    />)
+    .create(
+        <MockProvider>
+          <App places={mockData}/>)
+        </MockProvider>
+    )
     .toJSON();
 
   expect(tree).toMatchSnapshot();
