+++ src/components/cities-list/cities-list.test.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import renderer from 'react-test-renderer';
 import CitiesList from './cities-list';
 
@@ -11,12 +11,11 @@ const citiesMock = [
 
 it(`Cities List correctly renders`, () => {
   const tree = renderer
-    .create(
-      <CitiesList
-        cities={citiesMock}
-        city={`Paris`}
-        onCityClick={jest.fn()}
-      />)
+    .create(<CitiesList
+      cities={citiesMock}
+      city={`Paris`}
+      onCityClick={jest.fn()}
+    />)
     .toJSON();
   expect(tree).toMatchSnapshot();
 });
