+++ src/components/map/map.test.jsx
@@ -1,24 +1,9 @@
-import * as React from 'react';
+import React from 'react';
 import renderer from 'react-test-renderer';
-import App from './app';
-import {createStore} from 'redux';
-import {initialState, reducer} from '../../reducer';
-import {Provider} from 'react-redux';
-
-const createMockStore = (state = initialState) => createStore(reducer, state);
-
-export const MockProvider = ({state = initialState, children} = {}) => (
-  <Provider store={createMockStore(state)}>
-    {children}
-  </Provider>
-);
+import Map from './map';
 
 const mockData = [
   {
-    city: {
-      name: `Paris`,
-      coords: [1, 3],
-    },
     title: `Beautiful, luxurious apartment at great location`,
     type: `Apartment`,
     coords: [52.3909553943508, 4.85309666406198],
@@ -29,10 +14,6 @@ const mockData = [
     isPremium: true
   },
   {
-    city: {
-      name: `Amsterdam`,
-      coords: [48.8351, 2.3425],
-    },
     title: `Wood and stone place`,
     type: `Private room`,
     coords: [52.369553943508, 4.85309666406198],
@@ -43,10 +24,6 @@ const mockData = [
     isPremium: false
   },
   {
-    city: {
-      name: `Cologne`,
-      coords: [150, 2],
-    },
     title: `Canal View Prinsengracht`,
     type: `Apartment`,
     coords: [52.3909553943508, 4.929309666406198],
@@ -57,10 +34,6 @@ const mockData = [
     isPremium: false
   },
   {
-    city: {
-      name: `Berlin`,
-      coords: [11, 16],
-    },
     title: `Nice, cozy, warm big bed apartment`,
     type: `Apartment`,
     coords: [52.3809553943508, 4.939309666406198],
@@ -69,20 +42,6 @@ const mockData = [
     rate: 100,
     isBookmarked: false,
     isPremium: true
-  },
-  {
-    city: {
-      name: `Cologne`,
-      coords: [150, 2],
-    },
-    title: `Canal View Prinsengracht`,
-    type: `Apartment`,
-    coords: [52.3909553943508, 4.929309666406198],
-    image: `img/apartment-02.jpg`,
-    price: `132`,
-    rate: 80,
-    isBookmarked: true,
-    isPremium: false
   }
 ];
 
@@ -95,12 +54,10 @@ const fixContainerLeafletTest = () => {
 fixContainerLeafletTest();
 
 it(`App correctly renders`, () => {
+  Map.prototype.componentDidMount = () => {};
+
   const tree = renderer
-    .create(
-        <MockProvider>
-          <App places={mockData}/>)
-        </MockProvider>
-    )
+    .create(<Map offers={mockData} />)
     .toJSON();
 
   expect(tree).toMatchSnapshot();
