+++ src/components/place-card/place-card.e2e.test.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import PlaceCard from './place-card';
