+++ src/components/place-card/place-card.test.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import renderer from 'react-test-renderer';
 import PlaceCard from './place-card';
