+++ src/index.jsx
@@ -1,9 +1,9 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import React from 'react';
+import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {createStore, applyMiddleware} from 'redux';
-import {composeWithDevTools} from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
+import {compose} from 'recompose';
 import App from './components/app/app';
 import {offers} from './mocks/offers';
 import {reducer} from "./reducer";
@@ -12,9 +12,10 @@ import api from './api';
 
 const store = createStore(
     reducer,
-    composeWithDevTools(
-      applyMiddleware(thunk.withExtraArgument(api)),
-  ),
+    compose(
+        applyMiddleware(thunk.withExtraArgument(api)),
+        window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : (a) => a
+    )
 );
 
 ReactDOM.render(
