--- tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "sourceMap": true,
-    "module": "commonjs",
-    "target": "es2016",
-    "jsx": "react",
-    "allowSyntheticDefaultImports": false
-  },
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
-}
